refactor and fix filesystem

+++ ulib/ruxmusl/src/aarch64/mod.rs
@@ -84,7 +84,7 @@ pub fn syscall(syscall_id: SyscallId, args: [usize; 6]) -> isize {
             SyscallId::CHDIR => ruxos_posix_api::sys_chdir(args[0] as *const c_char) as _,
             #[cfg(feature = "fs")]
             SyscallId::OPENAT => ruxos_posix_api::sys_openat(
-                args[0],
+                args[0] as c_int,
                 args[1] as *const core::ffi::c_char,
                 args[2] as c_int,
                 args[3] as ctypes::mode_t,
